requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
colorama==0.4.6
distro==1.9.0
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
openai==1.97.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
python-bidi==0.6.6
python-docx==1.2.0
python-dotenv==1.1.1
python-telegram-bot==20.7
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.16.1
shapely==2.1.1
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
